requirements.txt
bio==1.3.3
gfapy==1.2.3
hydra-core==1.1
icontract==2.6.1
networkx==2.8
numpy
typeguard==2.13.3
